chromeos_gn.bbclass: add helper class
The `chromeos_gn` class sets up a build environment ready to compile `gn`
recipes which use `common-mk`.

By default, the `OUTPUT_DIR` is defined to be the same as `${B}`, but this
can be overriden if desired.

The actual call to `gn` passes in `--root='${WORKDIR}/src/platform2'` and
`--args='${GN_ARGS}'`. In your recipes you should add `GN_ARGS +=` for
recipe-specific arguments such as `platform_subdir` and the `use` flags list.

If you call `bitbake -D <recipe>` (first level of debugging), a list of valid
`--args` will be output in `log.do_configure`.

The only task defined in the class is `chromeos_gn_do_configure`, as
`do_compile` and `do_install` likely need recipe-specific tweaks.

Signed-off-by: Tim Orling <tim.orling@konsulko.com>