chromiumos-platform/*: add *.inc and gen script
For recipes which have sources in `platform` repositories.

Since the pattern of `SRC_URI`s for the `chromiumos/src/platform` repositories
is consistent and the sheer number of repositories warranted semi-automation,
a script `recipes-chromeos/chromiumos-platform/gen-platform-inc.sh` is available
to generate the `chromiumos-platform-<name>.inc` files to be used in `platform`
recipes. It should be run in the `recipes-chromeos/chromiumos-platform`
directory and takes the path to `chromiumos/.repo/manifests/full.xml` as an
argument.

**NOTE:** `chromiumos-platform-cr50.inc` is a special case that had to be
manually created, since it is in the `ec` platform repository, but the branch
name is `cr50_stab`. There are probably other exceptions not handled by the
script.

Signed-off-by: Tim Orling <tim.orling@konsulko.com>