gen-chromeos-base-recipes.sh: add helper script
A very naïve recipe generator script is available in
`scripts/gen-chromeos-base-recipes.sh`. It takes the path to
`chromiumos/src/third-party/chromeos-overlay/chromeos-base` as an argument.
It will delete and overwrite the `recipes-generated` directory structure with
simple skeleton recipes. It is not very intelligent, so recipes will need
changes. Generated recipes are created with a no-op `do_install()` task, so
this must be manually set by looking at `src_install` in the corresponding
[`*.ebuild`](https://chromium.googlesource.com/chromiumos/overlays/chromiumos-overlay/+/master/chromeos-base).The script currently does not know how to handle more than one `LICENSE`.
Generated recipes with two or more licenses will throw a warning during
parsing. This is because the syntax of the
[`LICENSE` variable](https://docs.yoctoproject.org/ref-manual/variables.html#term-LICENSE)
in Yocto Project uses [SPDX licenses](https://spdx.org/licenses/) and uses
`& | ( )` to indicate "AND", "OR" and compound logic).

Add .gitignore for recipes-generated as we do not want to commit
accidentally.

Signed-off-by: Tim Orling <tim.orling@konsulko.com>