chromeos-image.bbclass: add helper class
Similar to core-image.bbclass, this class can be inherited by image
recipes to enable ChromeOS specific features.

In particular, the CHROMEOS_IMAGE_EXTRA_INSTALL variable
"Specifies the list of packages to be added to an image which inherits the
chromeos-image class. You should only set this variable in the
conf/local.conf file in the Build Directory."

This is handy for development of new recipes for an existing image.

Signed-off-by: Tim Orling <tim.orling@konsulko.com>